refactor(extract_atoms): peel the supersede helper into a sibling module
Keeps extract-atoms.ts under the unlisted-file module-size cap without
touching the ratchet TSV.
